(REF) separate general surface plotting utils from globe-specific functionality
