+++ hypercoil/viz/globe.py
@@ -6,18 +6,23 @@ Brain globe plots.
 """
 import torch
 import numpy as np
-import nibabel as nb
 import templateflow.api as tflow
 import matplotlib.pyplot as plt
-from matplotlib.colors import ListedColormap
 from hypercoil.functional.sphere import sphere_to_latlong
 from hypercoil.neuro.const import fsLR
+from hypercoil.viz.surfutils import (
+    data_from_struc_tag,
+    _SurfFromFilesMixin,
+    _SurfNoActionMixin,
+    _CMapFromSurfMixin,
+    _CMapNoActionMixin
+)
 
 
-def brain_globe(data, coor, shift=0, cmap='flag',
-                figsize=(12, 12), projection='mollweide'):
+def brain_globe(data, coor, shift=0, cmap='flag', figsize=(12, 12),
+                projection='mollweide', dtype=None, device=None):
     if not isinstance(data, torch.Tensor):
-        data = torch.tensor(data)
+        data = torch.tensor(data, dtype=dtype, device=device)
     if not isinstance(coor, torch.Tensor):
         coor = torch.tensor(coor, dtype=data.dtype, device=data.device)
 
@@ -44,32 +49,20 @@ def brain_globe(data, coor, shift=0, cmap='flag',
     return ax
 
 
-def data_from_struc_tag(cifti, struc_tag, dtype=None, device=None):
-    if struc_tag is not None:
-        slices = []
-        brain_model_axis = cifti.header.get_axis(1)
-        for struc, slc, _ in brain_model_axis.iter_structures():
-            if struc in struc_tag:
-                if slc.stop is None:
-                    stop = cifti.shape[-1]
-                    slc = slice(slc.start, stop, slc.step)
-                slices.append(slc)
-        slices = np.r_[tuple(slices)]
-        data = cifti.get_fdata()[:, slices]
-    else:
-        data = cifti.get_fdata()
-    return torch.tensor(data, dtype=dtype, device=device)
-
-
 class _GlobeBrain:
     def __init__(self, data, coor=None, shift=0,
                  coor_mask=None, struc_tag=None, cmap='flag',
-                 figsize=(12, 12), projection='mollweide'):
+                 figsize=(12, 12), projection='mollweide',
+                 dtype=None, device=None):
         self.data, self.coor = self._select_data_and_coor(
             data, coor,
             coor_mask,
             struc_tag
         )
+        self.coor = self._transform_coor(
+            self.coor,
+            dtype=dtype,
+            device=device)
         self.cmap = self._select_cmap(cmap, struc_tag)
         self.shift = shift
         self.figsize = figsize
@@ -78,6 +71,11 @@ class _GlobeBrain:
     def drop_null(self, null=0):
         subset = (self.data != null)
         return self.data[subset], self.coor[subset]
+
+    def _transform_coor(self, coor, dtype, device):
+        coor = sphere_to_latlong(
+            torch.tensor(coor, dtype=dtype, device=device))
+        return coor
 
     def __call__(self):
         return brain_globe(
@@ -103,30 +101,6 @@ class _GlobeCortexR(_GlobeBrain):
             data=data, coor=coor, shift=(np.pi / 4),
             cmap=cmap, figsize=figsize, projection=projection
         )
-
-
-class _SurfNoActionMixin:
-    def _select_data_and_coor(self, data, coor,
-                              coor_mask=None,
-                              struc_tag=None):
-        return data, coor
-
-
-class _SurfFromFilesMixin:
-    def _select_data_and_coor(self, data, coor,
-                              coor_mask=None,
-                              struc_tag=None,
-                              dtype=None, device=None):
-        data = nb.load(data)
-        coor = nb.load(coor).darrays[0].data
-        if coor_mask is not None:
-            coor_mask = nb.load(coor_mask)
-            coor_mask = coor_mask.darrays[0].data.astype(bool)
-            coor = coor[coor_mask]
-        coor = sphere_to_latlong(
-            torch.tensor(coor, dtype=dtype, device=device))
-        data = data_from_struc_tag(data, struc_tag)
-        return data, coor
 
 
 class _CortexLfsLR32KMixin(_SurfFromFilesMixin):
@@ -165,59 +139,6 @@ class _CortexRfsLR32KMixin(_SurfFromFilesMixin):
             cmap=cmap, struc_tag='CIFTI_STRUCTURE_CORTEX_RIGHT')
 
 
-class _CMapNoActionMixin:
-    def _select_cmap(self, cmap, struc_tag=None):
-        return cmap
-
-
-class _CMapFromSurfMixin:
-    def _compute_linear_map(self, null=0, dtype=None, device=None):
-        labels = np.unique(self.data).astype(int)
-        labels = np.delete(labels, labels==null)
-        n_labels = labels.max() + 1
-        n_voxels = np.prod(self.data.shape[:3])
-        map = np.zeros((n_labels, n_voxels))
-        for l in labels:
-            map[l, :] = (self.data == l)
-        map /= map.sum(1, keepdims=True)
-        map[np.isnan(map)] = 0
-        return torch.tensor(map, dtype=dtype, device=device)
-
-    def _compute_parcel_colours(self, linear_map, surf_cmap):
-        parcel_colours = linear_map @ surf_cmap.T
-        parcel_colours = parcel_colours.numpy()
-        parcel_colours = np.maximum(parcel_colours, 0)
-        parcel_colours = np.minimum(parcel_colours, 1)
-        return parcel_colours
-
-    def _align_cmap_to_data(self, dataset, cmap):
-        null = 0 # hard coding this now
-        present_in_dset = np.unique(dataset).astype(int)
-        present_in_dset = np.delete(
-            present_in_dset,
-            present_in_dset==null
-        )
-        present_map = np.arange(len(present_in_dset)).astype(int) + 1
-        present_band = np.zeros(cmap.colors.shape[0]).astype(int)
-        present_band[present_in_dset] = present_map
-        # This shift sort of depends on where null is, so right now your
-        # null had better be zero
-        cmap = cmap.colors[present_in_dset]
-        dataset = dataset.int()
-        dataset = present_band[dataset]
-        return dataset, ListedColormap(cmap)
-
-    def _select_cmap(self, cmap, struc_tag=None):
-        surf_cmap = nb.load(cmap)
-        surf_cmap = data_from_struc_tag(surf_cmap, struc_tag)
-        linear_map = self._compute_linear_map(null=0)
-        parcel_colours = self._compute_parcel_colours(linear_map, surf_cmap)
-        cmap = ListedColormap(parcel_colours)
-        self.data, cmap = self._align_cmap_to_data(self.data, cmap)
-        self.data, self.coor = self.drop_null(null=0)
-        return cmap
-
-
 class _ModalGlobeMixin:
     def _config_map_and_cmap(self):
         pass
@@ -228,17 +149,33 @@ class _NetworkGlobeMixin:
         pass
 
 
-class GlobeBrain(_GlobeBrain, _SurfNoActionMixin, _CMapNoActionMixin):
+class GlobeBrain(
+    _GlobeBrain,
+    _SurfNoActionMixin,
+    _CMapNoActionMixin
+):
     pass
 
 
-class GlobeFromFiles(_GlobeBrain, _SurfFromFilesMixin, _CMapNoActionMixin):
+class GlobeFromFiles(
+    _GlobeBrain,
+    _SurfFromFilesMixin,
+    _CMapNoActionMixin
+):
     pass
 
 
-class CortexLfsLRFromFiles(_GlobeCortexL, _CortexLfsLR32KMixin, _CMapFromSurfMixin):
+class CortexLfsLRFromFiles(
+    _GlobeCortexL,
+    _CortexLfsLR32KMixin,
+    _CMapFromSurfMixin
+):
     pass
 
 
-class CortexRfsLRFromFiles(_GlobeCortexR, _CortexRfsLR32KMixin, _CMapFromSurfMixin):
+class CortexRfsLRFromFiles(
+    _GlobeCortexR,
+    _CortexRfsLR32KMixin,
+    _CMapFromSurfMixin
+):
     pass
